don't try to delete unsaved object

+++ custom/ilsgateway/tests/test_reminders2.py
@@ -46,7 +46,8 @@ class RemindersTest(ILSTestScript):
     @classmethod
     def tearDownClass(cls):
         delete_domain_phone_numbers(TEST_DOMAIN)
-        cls.sms_backend_mapping.delete()
+        if cls.sms_backend_mapping.id is not None:
+            cls.sms_backend_mapping.delete()
         cls.sms_backend.delete()
         cls.domain.delete()
         super(RemindersTest, cls).tearDownClass()
